Move split ledger test
