+++ tests/e2e_operations.py
@@ -14,7 +14,7 @@ import ipaddress
 import infra.interfaces
 import infra.path
 import infra.proc
-
+import random
 
 from loguru import logger as LOG
 
@@ -149,6 +149,59 @@ def test_forced_snapshot(network, args):
     raise RuntimeError("Could not find matching snapshot file")
 
 
+def split_all_ledger_files_in_dir(input_dir, output_dir):
+    # A ledger file can only be split at a seqno that contains a signature
+    # (so that all files end on a signature that verifies their integrity).
+    # We first detect all signature transactions in a ledger file and truncate
+    # at any one (but not the last one, which would have no effect) at random.
+    for ledger_file in os.listdir(input_dir):
+        sig_seqnos = []
+
+        if ledger_file.endswith(ccf.ledger.RECOVERY_FILE_SUFFIX):
+            # Ignore recovery files
+            continue
+
+        ledger_file_path = os.path.join(input_dir, ledger_file)
+        ledger_chunk = ccf.ledger.LedgerChunk(ledger_file_path, ledger_validator=None)
+        for transaction in ledger_chunk:
+            public_domain = transaction.get_public_domain()
+            if ccf.ledger.SIGNATURE_TX_TABLE_NAME in public_domain.get_tables().keys():
+                sig_seqnos.append(public_domain.get_seqno())
+
+        if len(sig_seqnos) <= 1:
+            # A chunk may not contain enough signatures to be worth truncating
+            continue
+
+        # Ignore last signature, which would result in a no-op split
+        split_seqno = random.choice(sig_seqnos[:-1])
+
+        assert ccf.split_ledger.run(
+            [ledger_file_path, str(split_seqno), f"--output-dir={output_dir}"]
+        ), f"Ledger file {ledger_file_path} was not split at {split_seqno}"
+        LOG.info(
+            f"Ledger file {ledger_file_path} was successfully split at {split_seqno}"
+        )
+        LOG.debug(f"Deleting input ledger file {ledger_file_path}")
+        os.remove(ledger_file_path)
+
+
+@reqs.description("Split ledger")
+def test_split_ledger_on_stopped_network(primary, args):
+    # Test that ledger files can be arbitrarily split.
+    # Note: For real operations, it would be best practice to use a separate
+    # output directory
+
+    current_ledger_dir, committed_ledger_dirs = primary.get_ledger()
+    split_all_ledger_files_in_dir(current_ledger_dir, current_ledger_dir)
+    if committed_ledger_dirs:
+        split_all_ledger_files_in_dir(
+            committed_ledger_dirs[0], committed_ledger_dirs[0]
+        )
+
+    # Check that the split ledger can be read successfully
+    ccf.ledger.Ledger([current_ledger_dir] + committed_ledger_dirs, False)
+
+
 def run_file_operations(args):
     with tempfile.TemporaryDirectory() as tmp_dir:
         txs = app.LoggingTxs("user0")
@@ -168,6 +221,11 @@ def run_file_operations(args):
             test_parse_snapshot_file(network, args)
             test_forced_ledger_chunk(network, args)
             test_forced_snapshot(network, args)
+
+            primary, _ = network.find_primary()
+            network.stop_all_nodes()
+
+            test_split_ledger_on_stopped_network(primary, args)
 
 
 def run_tls_san_checks(args):
